refactor: 용도에 맞게 패키지 구분
https://github.com/next-step/java-bowling/pull/957#issuecomment-1312388691
