+++ src/main/java/bowling/domain/Player.java
@@ -1,4 +1,7 @@
-package bowling;
+package bowling.domain;
+
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.NormalFrame;
 
 import java.util.ArrayList;
 import java.util.List;
