+++ src/test/java/bowling/ChancesTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.Chances;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
