+++ src/test/java/bowling/PlayerTest.java
@@ -1,5 +1,8 @@
 package bowling;
 
+import bowling.domain.frame.Frame;
+import bowling.domain.Player;
+import bowling.domain.PlayerDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
