+++ src/test/java/bowling/StatusTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.Status;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
